+++ src/types.nim
@@ -115,7 +115,7 @@ type
     video*: Option[Video]
 
   Quote* = object
-    id*: int
+    id*: int64
     profile*: Profile
     text*: string
     reply*: seq[string]
@@ -128,7 +128,7 @@ type
 
   Retweet* = object
     by*: string
-    id*: int
+    id*: int64
 
   TweetStats* = object
     replies*: string
@@ -136,8 +136,8 @@ type
     likes*: string
 
   Tweet* = ref object
-    id*: int
-    threadId*: int
+    id*: int64
+    threadId*: int64
     profile*: Profile
     text*: string
     time*: Time
@@ -159,7 +159,7 @@ type
 
   Chain* = ref object
     content*: seq[Tweet]
-    more*: int
+    more*: int64
 
   Conversation* = ref object
     tweet*: Tweet
